+++ haskell-lsp-types/ChangeLog.md
@@ -1,5 +1,24 @@
 # Revision history for haskell-lsp-types
 
+## 0.21.0.0
+
+* Stop getCompletionPrefix from crashing if beforePos is empty
+* Add DidChangeWatchedFilesRegistrationOptions
+* Add NormalizedFilePath from ghcide
+* Add diagnostic and completion tags
+* Fix language server example
+* Correctly fix the problem with '$/' notifications
+* Add azure ci
+
+## 0.20.0.0
+
+* Force utf8 encoding when writing vfs temp files
+* Don't log errors for '$/' notifications (@jinwoo)
+* Force utf8 encoding when writing vfs temp files (@jneira)
+* Store a hash in a NormalizedUri (@mpickering)
+* Move "Semigroup WorkspaceEdit" instance (@sheaf)
+* Fix vfs line endings (@jneira)
+
 ## 0.19.0.0 -- 2019-12-14
 
 * Fix vfs line endings (@jneira)